+++ release-notes.md
@@ -59,6 +59,7 @@
 - Fixed bug causing friendly/enemy teams to switch mid-demo when using `teamlock 1` (3.2 bug)
 - Fixed bug causing rendering issue when using `/gl_contrast` to brighten screen in classic renderer (3.5 bug, #442, reported by hammer)
 - Fixed bug causing no rendering of aliasmodels when VAs not supported but `/gl_program_aliasmodels` set (3.5 bug)
+- Fixed bug causing atlas textures to not be rendered when using ATI drivers reporting version x.y.13399 (seems to be driver issue? #416)
 
 ### Ruleset-related changes